Git it working again

+++ TextBasedGameWithPointers/Utilities.cpp
@@ -1,4 +1,6 @@
 #include "Utilities.h"
+#include <ctime>
+#include <cstdlib>    
 
 Utilities::Utilities()
 {
@@ -11,5 +13,5 @@ Utilities::~Utilities()
 int Utilities::chooseRandom(int numberOfOptions)
 {
 	srand(time(NULL));
-	return (rand() * 5009) % numberOfOptions;
+	return uint32_t(rand() *5009) % numberOfOptions;
 }
